🔥 Remove unused code
This fork of dbt loom makes use of the `get_plugin_artifacts` hook in
the dbt plugin api. As such, many of the methods related to parsing
external nodes are not required.

Additionally, this commit updates naming to ensure all resources are
appropriately named.
