+++ dbt_remote_state/clients/az_blob.py
@@ -7,7 +7,7 @@ from typing import Dict
 
 from pydantic import BaseModel
 
-from dbt_loom.logging import fire_event
+from dbt_remote_state.logging import fire_event
 
 
 class AzureReferenceConfig(BaseModel):
