+++ dbt_remote_state/manifests.py
@@ -1,96 +1,35 @@
-import datetime
 from io import BytesIO
 import json
 import gzip
 import os
 from pathlib import Path
-from typing import Dict, List, Optional
+from typing import Dict
 from urllib.parse import unquote, urlunparse
 
-from pydantic import BaseModel, Field, validator
 import requests
 
-from dbt_loom.clients.snowflake_stage import SnowflakeReferenceConfig, SnowflakeClient
+from dbt_remote_state.clients.snowflake_stage import (
+    SnowflakeReferenceConfig,
+    SnowflakeClient,
+)
 
 try:
     from dbt.artifacts.resources.types import NodeType
 except ModuleNotFoundError:
-    from dbt.node_types import NodeType  # type: ignore
+    pass  # type: ignore
 
-from dbt_loom.clients.az_blob import AzureClient, AzureReferenceConfig
-from dbt_loom.clients.dbt_cloud import DbtCloud, DbtCloudReferenceConfig
-from dbt_loom.clients.paradime import ParadimeClient, ParadimeReferenceConfig
-from dbt_loom.clients.gcs import GCSClient, GCSReferenceConfig
-from dbt_loom.clients.s3 import S3Client, S3ReferenceConfig
-from dbt_loom.clients.dbx import DatabricksClient, DatabricksReferenceConfig
-from dbt_loom.config import (
+from dbt_remote_state.clients.az_blob import AzureClient, AzureReferenceConfig
+from dbt_remote_state.clients.dbt_cloud import DbtCloud, DbtCloudReferenceConfig
+from dbt_remote_state.clients.paradime import ParadimeClient, ParadimeReferenceConfig
+from dbt_remote_state.clients.gcs import GCSClient, GCSReferenceConfig
+from dbt_remote_state.clients.s3 import S3Client, S3ReferenceConfig
+from dbt_remote_state.clients.dbx import DatabricksClient, DatabricksReferenceConfig
+from dbt_remote_state.config import (
     FileReferenceConfig,
-    LoomConfigurationError,
+    RemoteStateConfigurationError,
     ManifestReference,
     ManifestReferenceType,
 )
-
-
-class DependsOn(BaseModel):
-    """Wrapper for storing dependencies"""
-
-    nodes: List[str] = Field(default_factory=list)
-    macros: List[str] = Field(default_factory=list)
-
-
-class ManifestNode(BaseModel, use_enum_values=True):
-    """A basic ManifestNode that can be referenced across projects."""
-
-    name: str
-    package_name: str
-    unique_id: str
-    resource_type: NodeType
-    schema_name: str = Field(alias="schema")
-    database: Optional[str] = None
-    relation_name: Optional[str] = None
-    version: Optional[str] = None
-    latest_version: Optional[str] = None
-    deprecation_date: Optional[datetime.datetime] = None
-    access: Optional[str] = "protected"
-    group: Optional[str] = None
-    generated_at: datetime.datetime = Field(default_factory=datetime.datetime.utcnow)
-    depends_on: Optional[DependsOn] = None
-    depends_on_nodes: List[str] = Field(default_factory=list)
-    enabled: bool = True
-
-    @validator("depends_on_nodes", always=True)
-    def default_depends_on_nodes(cls, v, values):
-        depends_on = values.get("depends_on")
-        if depends_on is None:
-            return []
-
-        return [
-            node for node in depends_on.nodes if node.split(".")[0] not in ("source")
-        ]
-
-    @validator("resource_type", always=True)
-    def fix_resource_types(cls, v, values):
-        """If the resource type does not match the unique_id prefix, then rewrite the resource type."""
-
-        node_type = values.get("unique_id").split(".")[0]
-        if v != node_type:
-            return node_type
-        return v
-
-    @property
-    def identifier(self) -> str:
-        if not self.relation_name:
-            return self.name
-
-        return self.relation_name.split(".")[-1].replace('"', "").replace("`", "")
-
-    def dump(self) -> Dict:
-        """Dump the ManifestNode to a Dict, with support for pydantic 1 and 2"""
-        exclude_set = {"schema_name", "depends_on", "node_config", "unique_id"}
-        if hasattr(self, "model_dump"):
-            return self.model_dump(exclude=exclude_set)  # type: ignore
-
-        return self.dict(exclude=exclude_set)
 
 
 class UnknownManifestPathType(Exception):
@@ -111,7 +50,7 @@ class ManifestLoader:
             ManifestReferenceType.azure: self.load_from_azure,
             ManifestReferenceType.snowflake: self.load_from_snowflake,
             ManifestReferenceType.paradime: self.load_from_paradime,
-            ManifestReferenceType.databricks: self.load_from_databricks
+            ManifestReferenceType.databricks: self.load_from_databricks,
         }
 
     @staticmethod
@@ -148,7 +87,9 @@ class ManifestLoader:
             )
 
         if not file_path.exists():
-            raise LoomConfigurationError(f"The path `{file_path}` does not exist.")
+            raise RemoteStateConfigurationError(
+                f"The path `{file_path}` does not exist."
+            )
 
         if file_path.suffix == ".gz":
             with gzip.open(file_path, "rt") as file:
@@ -239,7 +180,7 @@ class ManifestLoader:
             command_index=config.command_index,
         )
         return paradime_client.load_manifest()
-    
+
     @staticmethod
     def load_from_databricks(config: DatabricksReferenceConfig) -> Dict:
         """Load a manifest dictionary from Databricks."""
@@ -250,7 +191,7 @@ class ManifestLoader:
         """Load a manifest dictionary based on a ManifestReference input."""
 
         if manifest_reference.type not in self.loading_functions:
-            raise LoomConfigurationError(
+            raise RemoteStateConfigurationError(
                 f"The manifest reference provided for {manifest_reference.name} does "
                 "not have a valid type."
             )
@@ -259,9 +200,9 @@ class ManifestLoader:
             manifest = self.loading_functions[manifest_reference.type](
                 manifest_reference.config
             )
-        except LoomConfigurationError as e:
+        except RemoteStateConfigurationError:
             if getattr(manifest_reference, "optional", False):
-                return None
+                return None  # type: ignore
             raise
 
         return manifest
