+++ pyproject.toml
@@ -1,13 +1,13 @@
 [tool.poetry]
-name = "dbt-loom"
-version = "0.9.0"
-description = "A dbt-core plugin to import public nodes in multi-project deployments."
-authors = ["Nicholas Yager <yager@nicholasyager.com>"]
+name = "dbt-remote-state"
+version = "0.1.0"
+description = "A dbt-core plugin for working with remote state."
+authors = ["Nicholas Yager <yager@nicholasyager.com>", "Morgan Kerle <mdgk95@proton.me>"]
 readme = "README.md"
-packages = [{ include = "dbt_loom" }]
+packages = [{ include = "dbt_remote_state" }]
 
 [tool.commitizen]
-version = "0.9.0"
+version = "0.1.0"
 version_files = ["pyproject.toml:^version"]
 
 [tool.poetry.dependencies]
@@ -19,7 +19,6 @@ boto3 = "^1.28.84"
 azure-storage-blob = "^12.19.0"
 azure-identity = "^1.15.0"
 types-pyyaml = "^6.0.12.12"
-types-networkx = "^3.2.1.20240313"
 paradime-io = "^4.11.0"
 
 [tool.poetry.group.dev.dependencies]
@@ -30,6 +29,7 @@ dbt-duckdb = ">=1.6.0,<1.10.0"
 duckdb = ">=0.8.0"
 pre-commit = "^3.6.0"
 mypy = "^1.8.0"
+types-pytz = "^2025.2.0.20250516"
 
 [tool.poetry.extras]
 snowflake = ["dbt-snowflake"]
